+++ assets/javascripts/kitten/components/form/field/components/label.js
@@ -7,6 +7,7 @@ import { Line } from '../../../layout/line'
 
 export class FieldLabel extends Component {
   static propTypes = {
+    link: PropTypes.node,
     tooltip: PropTypes.string,
     labelProps: PropTypes.object,
   }
@@ -17,10 +18,17 @@ export class FieldLabel extends Component {
   }
 
   render() {
-    const { children, tooltip, tooltipId, labelProps } = this.props
+    const {
+      children,
+      tooltip,
+      tooltipId,
+      labelProps,
+      link,
+      ...others
+    } = this.props
 
     return (
-      <Marger bottom="1.5">
+      <Marger bottom="1.5" {...others}>
         <Line style={{ lineHeight: 1 }}>
           <Line.Item>
             <Label {...labelProps} size="tiny">
@@ -33,6 +41,8 @@ export class FieldLabel extends Component {
               <Tooltip id={tooltipId}>{tooltip}</Tooltip>
             </Line.Item>
           )}
+
+          {link && <Line.Item>{link}</Line.Item>}
         </Line>
       </Marger>
     )